% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_metrics.R
\name{normalize_metrics}
\alias{normalize_metrics}
\title{Normalize Optimal Topic Metrics}
\usage{
normalize_metrics(values)
}
\arguments{
\item{values}{optimal topics results}
}
\description{
This function allows you to normalize the output from optimal topic analysis
}
\examples{
normalize_metrics()
}
\keyword{normalize}
