% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_topics.R
\name{optimal_topics}
\alias{optimal_topics}
\title{Determine Optimal Topics}
\usage{
optimal_topics(dtm)
}
\arguments{
\item{dtm}{document-term matrix}
}
\description{
This function allows you to analyze document term matrix for latent structures
}
\examples{
optimal_topics(dtm)
}
\keyword{lda}
\keyword{optimal}
\keyword{topics}
