% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimal_topics.R
\name{plot_optimal_topics}
\alias{plot_optimal_topics}
\title{Plots Optimal Topic Analysis from Normalized Optimal Topic Values}
\usage{
plot_optimal_topics(values)
}
\arguments{
\item{values}{normalized cao and deveaud optimal topics output}
}
\description{
This function allows you to to plot the normalized Cao and Deveaud normalized latent structure values
}
\examples{
plot_optimal_topics(optimal_topics_norm_values)
}
\keyword{optimal}
\keyword{plot}
\keyword{topics}
