% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topics.R
\name{plot_topics}
\alias{plot_topics}
\title{Plots Topic Models from Latent Dirichlet Allocation Model}
\usage{
plot_topics(lda)
}
\arguments{
\item{lda}{latent dirichlet allocation model}
}
\description{
This function allows you to analyze topic models from an Latent Dirichlet Allocation model
}
\examples{
plot_topics(lda)
}
\keyword{lda}
\keyword{model}
\keyword{topic}
