% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_topics.R
\name{rank_topics}
\alias{rank_topics}
\title{Creates Rank Order of Optimal Topics}
\usage{
rank_topics(values)
}
\arguments{
\item{values}{normalized optimal topic metric output}
}
\description{
This function allows you to analyze the Cao and Deveaud optimal topics by rank ordering the metrics
}
\examples{
rank_topics(values)
}
\keyword{optimal}
\keyword{rank}
\keyword{topics}
