% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequency.R
\name{term_frequency}
\alias{term_frequency}
\title{N-gram term frequency for AFICA contract data}
\usage{
term_frequency(data, n = 1)
}
\arguments{
\item{data}{data frame where text has column name "description"}

\item{n}{number of n-grams for frequency analysis (n=1 for single term frequency)}
}
\description{
This function allows you to analyze n-gram term frequency for AFICA contract descriptions
}
\examples{
term_frequency(data, 2)
}
\keyword{n-gram}
