% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydtm.R
\name{tidyDTM}
\alias{tidyDTM}
\title{Constructing DTM for AFICA data}
\usage{
tidyDTM(text.df, sparse)
}
\arguments{
\item{text.df}{text data}

\item{sparse}{sparsity for which to trim document term matrix}
}
\description{
This function allows you to analyze document term matrix for latent structures
}
\examples{
tidyDTM(data, 0.98)
}
\keyword{dtm}
